requirements.txt
numpy==1.23.5
pandas==2.0.3
pyspark==3.5.5
dash==3.0.2
dash_mantine_components==1.1.1
delta-spark==3.3.0
gpxpy==1.6.2
pyarrow==14.0.1
Pillow==11.1.0